+++ tools/docgen/docgen.cpp
@@ -1,4 +1,4 @@
-// Copyright Glen Knowles 2020 - 2022.
+// Copyright Glen Knowles 2020 - 2023.
 // Distributed under the Boost Software License, Version 1.0.
 //
 // docgen.cpp - docgen
@@ -15,7 +15,7 @@ using namespace Dim;
 *
 ***/
 
-const VersionInfo kVersion = { 1, 1, 2 };
+const VersionInfo kVersion = { 1, 1, 3 };
 
 
 /****************************************************************************